fd_agents/requirements.txt
# FD Portfolio Optimizer - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LLM & AI
groq==0.4.1
crewai==0.25.0
langchain==0.1.0
langchain-groq==0.1.0

# Data Processing
python-dotenv==1.0.0
aiofiles==23.2.1

# Utilities
requests==2.31.0
tenacity==8.2.3
pydantic-settings==2.1.0

# Development (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.1
# flake8==6.1.0
